+++ src/test/java/org/folio/rest/support/matchers/DomainEventAssertions.java
@@ -3,7 +3,6 @@ package org.folio.rest.support.matchers;
 import static io.vertx.core.MultiMap.caseInsensitiveMultiMap;
 import static java.util.UUID.fromString;
 import static java.util.concurrent.TimeUnit.SECONDS;
-import static org.awaitility.Awaitility.await;
 import static org.folio.kafka.KafkaHeaderUtils.kafkaHeadersToMap;
 import static org.folio.okapi.common.XOkapiHeaders.TENANT;
 import static org.folio.okapi.common.XOkapiHeaders.URL;
@@ -26,12 +25,14 @@ import static org.hamcrest.CoreMatchers.nullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.greaterThan;
 import static org.hamcrest.Matchers.notNullValue;
+import static org.hamcrest.Matchers.empty;
 import static org.junit.Assert.assertEquals;
 
 import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
-
+import org.awaitility.Awaitility;
+import org.awaitility.core.ConditionFactory;
 import io.vertx.core.MultiMap;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
@@ -40,6 +41,16 @@ import io.vertx.kafka.client.producer.KafkaHeader;
 
 public final class DomainEventAssertions {
   private DomainEventAssertions() { }
+
+  /**
+   * Awaitility.await() with a default timeout of 2 seconds.
+   *
+   * <p>1 second is too short:
+   * <a href="https://issues.folio.org/browse/MODINVSTOR-754">MODINVSTOR-754</a>
+   */
+  private static ConditionFactory await() {
+    return Awaitility.await().atMost(2, SECONDS);
+  }
 
   private static void assertCreateEvent(KafkaConsumerRecord<String, JsonObject> createEvent, JsonObject newRecord) {
     assertThat("Create event should be present", createEvent.value(), is(notNullValue()));
@@ -90,42 +101,38 @@ public final class DomainEventAssertions {
   }
 
   public static void assertNoUpdateEvent(String instanceId) {
-    await().atLeast(1, SECONDS);
-    await().atMost(1, SECONDS)
-      .until(() -> getInstanceEvents(instanceId).size(), greaterThan(0));
+    await()
+      .until(() -> getInstanceEvents(instanceId), is(not(empty())));
 
     final JsonObject updateMessage  = getLastInstanceEvent(instanceId).value();
     assertThat(updateMessage.getString("type"), not(is("UPDATE")));
   }
 
   public static void assertNoUpdateEventForHolding(String instanceId, String hrId) {
-    await().atLeast(1, SECONDS);
-    await().atMost(1, SECONDS)
-      .until(() -> getHoldingsEvents(instanceId, hrId).size(), greaterThan(0));
+    await()
+      .until(() -> getHoldingsEvents(instanceId, hrId), is(not(empty())));
 
     final JsonObject updateMessage  = getLastHoldingEvent(instanceId, hrId).value();
     assertThat(updateMessage.getString("type"), not(is("UPDATE")));
   }
 
   public static void assertNoRemoveEvent(String instanceId) {
-    await().atLeast(1, SECONDS);
-    await().atMost(1, SECONDS)
-      .until(() -> getInstanceEvents(instanceId).size(), greaterThan(0));
+    await()
+      .until(() -> getInstanceEvents(instanceId), is(not(empty())));
 
     final JsonObject updateMessage  = getLastInstanceEvent(instanceId).value();
     assertThat(updateMessage.getString("type"), not(is("DELETE")));
   }
 
-  public static void assertNoCreateEvent(String instanceId) {
-    await().atLeast(1, SECONDS);
-
-    assertThat(getInstanceEvents(instanceId).size(), is(0));
+  public static void assertNoEvent(String instanceId) {
+    await().during(1, SECONDS)
+      .until(() -> getInstanceEvents(instanceId), is(empty()));
   }
 
   public static void assertCreateEventForInstance(JsonObject instance) {
     final String instanceId = instance.getString("id");
 
-    await().atMost(1, SECONDS)
+    await()
       .until(() -> getInstanceEvents(instanceId).size(), greaterThan(0));
 
     assertCreateEvent(getFirstInstanceEvent(instanceId), instance);
@@ -147,14 +154,14 @@ public final class DomainEventAssertions {
   public static void assertRemoveEventForInstance(JsonObject instance) {
     final String instanceId = instance.getString("id");
 
-    await().atMost(1, SECONDS)
+    await()
       .until(() -> getInstanceEvents(instanceId).size(), greaterThan(1));
 
     assertRemoveEvent(getLastInstanceEvent(instanceId), instance);
   }
 
   public static void assertRemoveAllEventForInstance() {
-    await().atMost(1, SECONDS)
+    await()
       .until(() -> getInstanceEvents(NULL_INSTANCE_ID).size(), greaterThan(0));
 
     assertRemoveAllEvent(getLastInstanceEvent(NULL_INSTANCE_ID));
@@ -163,7 +170,7 @@ public final class DomainEventAssertions {
   public static void assertUpdateEventForInstance(JsonObject oldInstance, JsonObject newInstance) {
     final String instanceId = oldInstance.getString("id");
 
-    await().atMost(1, SECONDS)
+    await()
       .until(() -> getInstanceEvents(instanceId).size(), greaterThan(1));
 
     assertUpdateEvent(getLastInstanceEvent(instanceId), oldInstance, newInstance);
@@ -173,7 +180,7 @@ public final class DomainEventAssertions {
     final String itemId = item.getString("id");
     final String instanceIdForItem = getInstanceIdForItem(item);
 
-    await().atMost(1, SECONDS)
+    await()
       .until(() -> getItemEvents(instanceIdForItem, itemId).size(), greaterThan(0));
 
     // Domain event for item has an extra 'instanceId' property for
@@ -187,7 +194,7 @@ public final class DomainEventAssertions {
     final String itemId = item.getString("id");
     final String instanceIdForItem = getInstanceIdForItem(item);
 
-    await().atMost(1, SECONDS)
+    await()
       .until(() -> getItemEvents(instanceIdForItem, itemId).size(), greaterThan(1));
 
     // Domain event for item has an extra 'instanceId' property for
@@ -198,7 +205,7 @@ public final class DomainEventAssertions {
   }
 
   public static void assertRemoveAllEventForItem() {
-    await().atMost(1, SECONDS)
+    await()
       .until(() -> getItemEvents(NULL_INSTANCE_ID, null).size(), greaterThan(0));
 
     assertRemoveAllEvent(getLastItemEvent(NULL_INSTANCE_ID, null));
@@ -208,7 +215,7 @@ public final class DomainEventAssertions {
     final String itemId = newItem.getString("id");
     final String instanceIdForItem = getInstanceIdForItem(newItem);
 
-    await().atMost(1, SECONDS)
+    await()
       .until(() -> getItemEvents(instanceIdForItem, itemId).size(), greaterThan(1));
 
     // Domain event for item has an extra 'instanceId' property for
@@ -223,7 +230,7 @@ public final class DomainEventAssertions {
     final String id = hr.getString("id");
     final String instanceId = hr.getString("instanceId");
 
-    await().atMost(1, SECONDS)
+    await()
       .until(() -> getHoldingsEvents(instanceId, id).size(), greaterThan(0));
 
     assertCreateEvent(getFirstHoldingEvent(instanceId, id), hr);
@@ -233,14 +240,14 @@ public final class DomainEventAssertions {
     final String id = hr.getString("id");
     final String instanceId = hr.getString("instanceId");
 
-    await().atMost(1, SECONDS)
+    await()
       .until(() -> getHoldingsEvents(instanceId, id).size(), greaterThan(1));
 
     assertRemoveEvent(getLastHoldingEvent(instanceId, id), hr);
   }
 
   public static void assertRemoveAllEventForHolding() {
-    await().atMost(1, SECONDS)
+    await()
       .until(() -> getHoldingsEvents(NULL_INSTANCE_ID, null).size(), greaterThan(0));
 
     assertRemoveAllEvent(getLastHoldingEvent(NULL_INSTANCE_ID, null));
@@ -250,7 +257,7 @@ public final class DomainEventAssertions {
     final String id = newHr.getString("id");
     final String instanceId = newHr.getString("instanceId");
 
-    await().atMost(1, SECONDS)
+    await()
       .until(() -> getHoldingsEvents(instanceId, id).size(), greaterThan(1));
 
     assertUpdateEvent(getLastHoldingEvent(instanceId, id), oldHr, newHr);